+++ src/components/Directory/Directory.jsx
@@ -9,34 +9,9 @@ import EmptyState from "./EmptyState";
 import Folder from "../Folder";
 import NewFolder from "../Popups/NewFolder";
 import NewFile from "../Popups/NewFile";
+import FileFilters from "./FileFilters";
 
 import css from "./Directory.module.scss";
-
-/* Filters files based on type */
-const FileFilters = ({ fileTypes = [], onFilterChange }) => {
-  const [selectedType, selectType] = useState("");
-
-  function setFilter(filter) {
-    selectType(filter);
-    onFilterChange(filter);
-  }
-
-  return (
-    <div className={css.fileFilters}>
-      {fileTypes.map((filter) => (
-        <div
-          key={filter}
-          className={`${css.fileFilter} ${
-            filter === selectedType && css.filterSelected
-          }`}
-          onClick={() => setFilter(filter)}
-        >
-          {filter.length !== 0 ? filter : "All"}
-        </div>
-      ))}
-    </div>
-  );
-};
 
 /* Content component for Directory */
 const Content = ({ files, folders, fileTypes, onFilterChange }) => {
