+++ src/components/Directory/FileFilters/index.js
@@ -1,0 +1,1 @@
+export { default } from "./FileFilters";
