+++ src/components/Popups/NewFile/NewFile.module.scss
@@ -11,12 +11,12 @@
   text-align: center;
   padding: 10px;
   border-radius: 5px;
-  background-color: $primary;
+  background-color: $primary-light;
   cursor: pointer;
   color: white;
 
   &:hover {
-    background-color: $primary-light;
+    background-color: $primary-light-1;
   }
 
   &:focus {
